peers_ch: use the new `peers_base.get_claim_summary` in `print_p_lines`
Instead of iterating over the `streams_info` dictionaries manually,
and building the line to be printed, we just call the method
`peers_base.get_claim_summary` with the arguments passed
from `print_p_lines`.

This method builds each line of information about the peers of each
claim. Then the claim number is added to the beginning of the line,
and stored in the `out` list.

Finally, the entire list is printed by `funcs.print_content`.