peers_ch: use the new methods in the peer channel methods
In `search_ch_peers` we remove the specific code
and use the `peers_base.search_m_claim_peers` method.

We add the name of the channel to the returning `peers_info`
dictionary.

----

We don't need a specific method to print the claim information
of the peer search.
Therefore, we remove the existing function `print_p_lines`
and just use the `peers_base.print_claims_lines` method.

This method allows us to print the claims as a paragraph,
or as individual lines, so we also have a new parameter `inline`
to control this.