peers_base: new method `search_m_claim_peers` to search peers
Given a list of strings, `claims`, with the URI or claim ID
of multiple claims, we can search the peers in these claims.

By default, with `resolve=True`, `claims` is expected
to be a list of strings.

But if `resolve=False` we assume `claims` is a list of already
resolved claims, that is, dictionaries obtained by first searching
the claims online.

Threads are used with the sub function `search_peers_th`
to search multiple claim for peers in parallel.
By default a maximum of 32 threads will be used.

----

The output of the search is a big dictionary with many keys:
- `'n_claims'`
- `'n_streams'`
- `'streams_info'`
- `'total_size'`
- `'total_duration'`
- `'streams_with_hosts'`
- `'streams_with_hosts_all'`
- `'total_peers'`
- `'total_peers_all'`
- `'unique_nodes'`
- `'unique_trackers'`
- `'peer_ratio'`
- `'peer_ratio_all'`
- `'hosting_coverage'`
- `'hosting_coverage_all'`
- `'local_node'`

----

The method uses `process_claims_peers` with `channel=False`,
so it doesn't add a `'channel'` key.