peers_base: new method `print_claims_lines` to print lines
Given a dictionary with fully processed information, `peers_info`,
which is the result of the peer search of multiple claims,
we can print the information for each claim to the terminal or to the file.

If `inline=True`, the default, a single line of information will be printed
for each claim.

If `inline=False`, then a full paragraph of information will be printed
for each claim.

This uses `get_claim_summary` to get the text information of each claim,
then we add the appropriate number before the line for final displaying.

We also handle the case in which the claim information does not exist,
so it cannot be passed to `get_claim_summary` successfully.