peers_claims: options to pass a claim directly without resolving
In `list_peers` we can pass a `claim` argument with the already
resolved claim, a dictionary, instead of using `uri`, `cid`, or `name`.

In `list_m_peers` we can set `resolve=False` so that the list `claims`
will be treated as containing already resolved claims, dictionaries,
and not just strings with URIs or claim IDs.

We update the corresponding documentation in `lbrytools.md`.