peers: print the lines regardless of whether there are streams or not
In the methods to find peers of multiple claims `peers_claims.list_m_peers`,
peers of a single channel `peers_ch.list_ch_peers`,
or peers of multiple channels `peers_multi.list_chs_peers`
we used to test for the existence of peers by looking at the returned peer information
of the search.

We would only print the resulting lines if we had at least one peer.
```
if n_streams > 0:
    print(...)
```

Now we don't do this test and we just print the lines, even if there are no streams,
and thus no peers.

This makes sure that an output is produced, even if there were no peers.

In the past we used this test because searching for peers was time consuming;
now that we have restructured the code, we don't need to do this test any more
because by the time we reach this part the time consuming part was already done,
and printing is very fast.