claims: separate the printing from `w_claim_search`
Instead of printing inside `w_claim_search`, this method will only return
a list of dictionaries with the claims found.

Then the list of claims are passed to `print_claims.print_tr_claims`
for final display.

The return value of `print_trending_claims` and `print_search_claims`
also changes to the output of `w_claim_search`.
The docstring is updated accordingly.