print_claims: `print_tr_claims` show the `'canonical_url'` of the claim
New option to show claim `'title'` instead of `'name'`, but only
if the `'title'` key actually exists; otherwise, the existing `'name'`
will be shown.

The return value is no longer the printed `content`. These methods
that print information don't typically return values,
they just print the processed input to the terminal or to a file.