claims: new method `params_claim_search` to parse the options
To search for trending claims we call the `claim_search`
command of `lbrynet` with the appropriate parameters and a page number.

We only need to establish the parameters once; the page number
is the only thing that will change between invocations.

So the new method `params_claim_search` establishes the parameters
in the appropriate form, whether we search for trending claims
or a specific text or tag.