claims: new method `claims_search_pg` to perform the search of one page
Inside the `w_claim_search` we use the `params_claim_search` method
to collect the options that will be passed to the `lbrynet claim search`
method.

Then we use the new `claim_search_pg` to perform the actual search.

If `page=None` (default) we loop from 1 to 20 to search the 20 pages
that `lbrynet claim search` can search normally.

If `page` is any other integer from 1 to 20, we only search
that particular page.