search_utils: expand output of `sort_filter_size`
This method first uses `sort_and_filter` to get the unique claims
passed to it, sorted by time, and then it uses `downloadable_size`
to calculate the total size and duration of the claims.

Therefore the output of `sort_filter_size` will combine the outputs
of `sort_and_filter`, the `'claims'` key, and the output
of `downloadable_size`, eight additonal keys:
- `'claims'`: list of dictionaries
- `'size'`: total size in bytes
- `'duration'`: total duration in seconds
- `'size_GB'`: total size converted to GiB
- `'d_h'`: integer part of the duration in hours
- `'d_min'`: integer part of the duration in minutes
- `'d_s'`: integer part of the duration in seconds
- `'days'`: total seconds converted to days
- `'summary'`: summary text including claims, size and duration

----

Add a new parameter `print_msg` to optionally print the summary
computed by `downloadable_size`.

It defaults to `False` so no summary is printed.