search_ch: `ch_search_fifty_claims` returns the full dictionary
The output of `search_utils.sort_filter_size` was expanded
to include additional output from `search_utils.downloadable_size`.

This dictionary is returned in full by `ch_search_fifty_claims`.

Then in the upper method `ch_search_latest`, we only return
the `'claims'` key, which is a list of claims, which may be empty
as well.