claims: calculate the size and duration of the searched claims
In `w_claim_search` after getting the list of claims we pass them
to `search_utils.downloadable_size` in order to obtain a dictionary
with the size and duration of all claims, and also
a paragraph of information that we can print.

We return from `w_claims_search` a dictionary that includes
the list of claims:
- `'claims'`: list of dictionaries
- `'searched'`: paragraph of text indicating what was searched

And also the return value of `search_utils.downloadable_size`:
- `'size'`: total size in bytes
- `'duration'`: total duration in seconds
- `'size_GB'`: total size converted to GiB
- `'d_h'`: integer part of the duration in hours
- `'d_min'`: integer part of the duration in minutes
- `'d_s'`: integer part of the duration in seconds
- `'days'`: total seconds converted to days
- `'summary'`: summary text including claims, size and duration

In `w_claim_search` we don't print the search parameters,
instead we print the summary at the end of the calling methods
`print_trending_claims` and `print_search_claims`.