claims: new `release` argument to `list_trending_claims`
It defaults to `False`.
If it is `True` it will print the release time of the claim,
if it exists (for stream claims like videos and audio files),
or the creation time, which all claims should have.

This argument is also added to `list_search_claims`.

The argument is passed to `print_claims.print_tr_claims`.

Also update the `lbrytools.md` documentation.