search_utils: rename the `'text'` key to `'summary'`
The `'text'` key in the return dictionary of `search_utils.downloadable_size`
is renamed to `'summary'`, as it's a summary of the claims processed,
including size and duration. Also update the docstring.

Since this method is used in `sort.sort_items_size`, the docstring
is updated as well.

The `sort.sort_items_size` method is also used in `printf.print_summary`
so the key used is changed from `'text'` to `'summary'`
and the docstring is also updated.