peers_ch: fix the case if a select channel doesn't exist or has no claims
Previously, if the channel is not found or has no claims
the `list_ch_peers` method would fail because printing
the lines of peer information with `peers_base.print_claims_lines`
would try accessing some data inside the `streams_info` key
of the `peers_info` dictionary, and these values would not be
available.

So now we just call `peers_base.print_claims_lines`
if there is at least one claim, with a resolved `stream`,
to guarante that the method will not fail.