funcs: define variables for time formatting
New variables define text formatting for using with `time.strftime`.

Format for displaying the full time in messages in the terminal.
It includes a space and the day of the week:
`TFMT = "%Y-%m-%d_%H:%M:%S%z %A"`

Format for displaying time that can be displayed in the terminal
or printed to a file. It has no spaces so it a single element:
`TFMTp = "%Y-%m-%d_%H:%M:%S%z"`

Format for creating file names with a time prefixed.
It has no spaces:
`TFMTf = "%Y%m%d_%H%M"`

----

We use `time.gmtime` instead of `time.localtime`
so that the displayed time is the same in every system.