playlists: split the `list_playlists` method into two methods
First a method `search_playlists` to get the playlists
from the preferences (wallet), and another method `print_playlists`
to only print the result of the search.

Now the return value of `list_playlists` will be the dictionary returned
by `search_playlists`.

The `list_playlists` method now has two new parameters `claim_id`
and `sanitize` to control whether to display the claim ID
of the claim in the playlists, and to remove the emojis
in the displayed name, if any.

----

Add a third method `print_summary_playlists` to display a summary
of the playlists, and the items inside of them.

This is called at the end of `list_playlists` after the claims
are printed with `print_playlists`.

----

We use the `TFMT` formats from the `funcs.py` module
to format the time string inside `print_playlists`.

We use `time.gmtime` instead of `time.localtime`
so that the displayed time is the same in every system.