parse: move `parse_claim_file` to a new module `parse.py`
This method extracts the claim ID of a list of claims, such as that
produced by `printf.print_summary`.

This can be used to process multiple claims either to download them
or to delete them, which is the reason parsing of a file is also
available from within the `download_multi.download_claims`
and `clean_multi.remove_claims` methods.

The `parse_claim_file` method is removed from `search.py`
and the `__init__.py` is adjusted.