printf: use the variables to format the time printing
We use the `TFMT` and `TFMTp` formats from the `funcs.py` module
to format the time string inside `print_f_claims`
and `print_summary`.

If we print a single line in a list of claims,
we use the "print" format `TFMTp` that has no spaces.

If we print a standalone time we use the regular format `TFMT`
that includes the day of the week after a space.

We use `time.gmtime` instead of `time.localtime`
so that the displayed time is the same in every system.