print: use the `funcs` variables to format the time printing
We use the `TFMT` format from the `funcs.py` module
to format the time string inside `print_info_pre_get`.

We use `time.gmtime` instead of `time.localtime`
so that the displayed time is the same in every system.

----

Rename `item` to `claim` because `print_info_pre_get`
should work with any claim.

We remove the return information of `print_info_pre_get`
because this method only needs to print information.