channels: use the `funcs` variables to format the time printing
We use the `TFMT` format from the `funcs.py` module
to format the time string inside `search_ch_subs_latest`.

We use the `TFMTp` format in the `print_ch_subs_latest` module.
If the `release_time` is not available, we display the `creation_timestamp`.

We use `time.gmtime` instead of `time.localtime`
so that the displayed time is the same in every system.

----

Use `value = claim["value"]` to shorten the code.

If `'release_time'` is not available, use `'creation_timestamp'`.