peers_base: use the `funcs` variables to format the time printing
We use the `TFMT` and `TFMTp` formats from the `funcs.py` module
to format the time string inside `get_claim_summary`.

If we get the summary in a single line, we use the "print" format `TFMTp`
that has no spaces.

If we get the summary as a paragraph, we use the regular format `TFMT`
that includes the day of the week after a space.

These formats are used for both the `creation_timestamp`
and the `release_time`.

We use `time.gmtime` instead of `time.localtime`
so that the displayed time is the same in every system.