blobs_ratio: use the `funcs` variables to format the time printing
We use the `TFMT` format from the `funcs.py` module
to format the time string inside `plot_histogram` and `print_blobs_ratio`.

We use `time.gmtime` instead of `time.localtime`
so that the displayed time is the same in every system.

----

The seeding `ratio` was adjusted to handle a division by zero.

Originally the `ratio` was set to 0.0 by default, and then
it would be calculated from the operation `blobs_up/blobs_down`.
If there was a division by zero, meaning downloaded blobs is zero,
the ratio would remain zero.

Now we check if the denominator is zero. If both deonominator
and numerator are zero, we set the ratio to `Indeter.`.
If only the denominator is zero we set the ratio to `Infinity`.

This is done both in the `plot_histogram` and `print_blobs_ratio`
methods.