comments: rename and move comment methods
The original `comments.py` module is removed, and its methods
are placed in three new modules.

We put basic methods for comments in the new module `comments_base.py`:
- `jsonrpc_post`: to send the message to the comment server
- `sign_comment`: to sign text messages with the channel's private key

We put the methods to list comments in `comments_list.py`:
- `augment_replies`
- `find_replies`
- `print_r_comments`
- `print_f_comments`
- `list_comments`: list comments from a particular publication;
  it's the main method that will be exposed at the top `lbrytools`
  namespace

We put methods to create comments in `comments_act.py`:
- `print_cmnt_result`
- `create_comment`: create a new comment or reply under a publication
- `get_ch_and_sign`
- `update_comment`: edit and update an already existing comment
- `abandon_comment`: remove an already existing comment
- `hide_comment`:  prototype code, currenty it's not used

The `__init__.py` is adjusted appropriately.