comments: use the `funcs` variables to format the time printing
In `comments_act.py` we use the `TFMT` format from the `funcs.py` module
to format the time string inside `print_cmnt_result` and `create_comment`.

The variable `cl_time` is renamed to `rels_time`,
and it uses `creation_timestamp` if `release_time` is not available.

Instead of printing each line immediately we collect the lines
in the `out` list, and then print the entire list.

----

In `comments_list.py` we use the `TFMTf` format for `print_f_comments`
because this method opens a file with the date as prefix optionally,
so the date must not have spaces.

We use `TFMT` format in `list_comments`.
The variable `cl_time` is renamed to `rels_time`,
and it uses `creation_timestamp` if `release_time` is not available.

Instead of printing each line immediately we collect the lines
in the `out` list, and then print the entire list.

----

We use `time.gmtime` instead of `time.localtime`
so that the displayed time is the same in every system.