print_claims: use the `funcs` variables to format the time printing
We use the `TFMTp` format from the `funcs.py` module
to format the time strings inside `print_tr_claims`
and `print_sch_claims`.

The `TFMTp` format is used to print information in a single line,
so it shows the time without spaces.

We use `time.gmtime` instead of `time.localtime`
so that the displayed time is the same in every system.