print_claims: new methods `get_fields` and `get_line`
The first new method `get_fields` method collects the information
from a claim and returns a dictionary and returns a dictionary
where each value is a string that can be printed in a line.

The parameter `long_chan` allows displaying the channel name as is (`False`),
or padding it with whitespace (`True`), so that many subsequent lines
are aligned.

The parameter `title` allows choosing to display the claim `'name'`
or claim `'title'`, if it exists.

The parameter `sanitize` allows removing the emojis from
the claim `'name'` or `'title'`, and the `'channel'` strings.
This is necessary in some interface systems that can't display emojis
properly.

----

The new `get_line` method first uses `get_field` to get the fields,
and then builds the text line that will be printed based on
the switches that are passed to it, `long_chan`, `create`, `height`, `release`,
`claim_id`, `typ`, `ch_name`, `sizes`, `fees`, `title`, and `sanitize`.

----

The `print_sch_claims` method is restructured to avoid calculating
the text fields. It cycles over the list of claims,
calls `get_line` on each claim, and builds a `line` variable
that is added to the list of lines `out`, which is finally printed
by `funcs.print_content` at the end of the method.

The `print_tr_claims` method is just the same `print_sch_claims`
but sets `ch_name=True` and `long_chan=True` by default.