search: use `uri` as `name` if `offline=True`
In `search_item` with `offline=True` the input can only be `cid` or `name`.
If `uri` is used, and not `name`, `uri` will be used for the `name` variable.

This guarantees that the `offline` option works, when the claim's `uri`
has already expired.

When `offline=False`, it will search for the claim online as usual.

The docstrings are updated to clarify that the `repost` parameter
only works with `offline=False`, as a repost must always be resolved
online.