print: update `print_info_pre_get` with new values
The `print.print_info_pre_get` method is updated to display
three additional values, the address of the publishing wallet,
the staked amount by the publisher, and the effective amount,
which includes the original staked amount plus all additional supports.

Now `print_info_pre_get` has an additional parameter `print_text`
which controls whether to actually display the information or not;
it defaults to `True`.

Now the return value is the paragraph of text calculated by the method,
which can be printed directly to the terminal or displayed in another interface.

This `print_info_pre_get` method already handles the `offline` parameter
passed from `check`.