claims_check: new `check` method to quickly display claim information
A new module `claims_check.py` is introduced with a single method, `check`.
This method can be used to quickly see the creation date of claims
or channels, the claim ID and claim title, and information
on the content, such as the release date, file type, duration,
and size, if available.

This method searches a claim with `search.search_item`,
and if found, it prints information about it
with `print.print_info_pre_get`.
The parameter `print_text` controls whether `print_info_pre_get`
actually prints the information or just returns the information.
The claim can be searched by canonical URL, claim ID, or name.

By default, if the claim is a repost, the information
of the original claim will be printed.
With `repost=False` the information of the repost claim itself
will be displayed.

By default the claim is searched online.
If we want to inspect a claim offline, that is,
in the list of downloaded claims, we can use `offline=True`.
In this case we must search by claim ID or name only;
if a URI is given, it will be used as the name.

Using `offline=True` has to be done for 'invalid' claims, that is,
claims which have been removed from the online database
but may still exist locally, if we previously downloaded them.

----

Update `__init__.py` so that `check` is in the main `lbrytool` namespace.

Also update the documentation `README.md` and `lbrytools.md`.