download: call `print_info_pre_get` after `search_item`
Inside the methods `download_collection`, `download_single`,
and `download_invalid` claims are resolved by calling
`search.search_item` on an address (URI, claim ID, or name).

The code is reorganized so that immediately after finding the claim
we print the information in the terminal using the extended
method `print.print_info_pre_get`.

The variable holding the claim is also renamed from `item` to `claim`
to clarify that it refers to a found claim.