download: new method `get_channel` to use as download directory
Inside `download_single` once we get the channel name from the given claim,
it is used as the name of a folder in which the downloaded file is placed.

For collections, we pass the original `ddir` to `download_collection`
through the `orig_ddir` variable.
Then inside `download_collection` we also extract the channel
of each claim to use as the containing folder.

The `own_dir` variable is also passed from `download_single`
to `download_collection`.