download: only create the directory if the claim is a `'stream'`
The `download_single` and `download_collection` methods try creating
a new directory in the system to place the contents of the claim,
if `own_dir=True`.

In the past, it would always try creating the new directory.
Now it will only create the directory if the claim is of type `'stream'`,
that is, if it contains a file that can be downloaded.

This means that trying to download other claims, like a `'channel'`,
or a `'repost'` when `repost=False`, or a `'collection'` when `collection=False`
won't create an empty directory unnecessarily.