funcs: methods to get the default download directory
Two new methods `default_ddir` and `get_download_dir`
that deal with the download directory.

The first method gets the download directory saved by the default
`lbrynet` configuration, or the home directory `~` if the server
is down.

The second method checks if the passed directory exists in the file system;
if it does exist, it simply returns it; if it doesn't,
it returns the default directory by calling `default_ddir`.