subscribed: handle channels that don't exist and those with no claims
In `ch_search_latest_d` we only need to search for claims in a channel
if this channel actually exists, meaning that it was resolved by calling `search_ch_subs`.

A new `'valid'` key is added to the output so that we can know if the channel
was resolved or not.

Then in `print_ch_subs_latest` the `'valid'` key is inspected together with `'claims'`.
It will print a different message depending on whether the channel
doesn't exist, that is, `'valid'` is `False`, or whether it exists but has no claims,
that is, `'claims'` is empty.