subscribed: rename `channels.py` to `subscribed.py`
The methods in this module refer to the channels in our subscriptions,
which are defined in our wallet. So the module is renamed
to reflect that "subscriptions" are the important part.

The modules that use the methods in this renamed module
are also updated, inlcluding `peers_multi.py` and the `__init__.py`.