blobs: pass `name` to the internal `c_blobs_th` method
Inside the internal method `c_blobs_th` we call `c_blobs` to find the
blobs of the given claim by claim ID or by name. If both are given
claim ID takes precedence but the name is also displayed
in case of an error.

This is used inside `count_blobs_all` which takes a list of claims
returned by `sort.sort_items`.

Also rename the variable `item` to `claim`.