blobs: update error message printing if the claims don't exist
When calling `count_blobs` or `count_blobs_all`,
if the claim is not found the internal `c_blobs` will return a dictionary
with the `'error_not_found'` key.

When using `search.search_item` we explicitly indicate `repost=True`
and `offline=False` as we want to know if an actual online claim
was downloaded. We also set `print_error=False`, and remove an empty `print()`
to avoid printing empty lines.

If the claim does exist but the `sd_hash` blob isn't in the system,
because it hasn't been downloaded, then it will return the `'error_no_sd_hash'` key
in the dictionary instead.

Then inside `prnt_blobs`, if any of these two keys is found,
it will display the value of the key as the error message and return.